修改数据类型

+++ legou-order/legou-order-service/src/main/java/com/hh/legou/order/controller/CartController.java
@@ -38,7 +38,7 @@ public class CartController {
      * @return
      */
     @RequestMapping("/add")
-    public ResponseEntity add(Long id, Integer num) throws IOException {
+    public ResponseEntity add(Long id, Long num) throws IOException {
         //从 Spring security获取当前用户
         Map<String, String> userInfo = tokenDecode.getUserInfo();
         String username = userInfo.get("user_name");
